requirements.txt
google-genai
python-dotenv
aiohttp
beautifulsoup4
requests
pytz
lxml
feedparser
yfinance
